Set relationship to CCMS to be from web
When viewing the container diagram it would only show one relationship
with CCMS, omitting the second one.

To fix this, we set the system relationship to CCMS to be from the
container that owns that relationship, instead of the system as a
whole.

This way the relationship is correctly summarised on all diagrams, but
still shows all the detail on the CCMS side as well.