Move internal relationships to a separate method
This commit puts all internal relationship declaration in a new method
called defineInteralContainerRelationships.

This is to separate the definition of internal containers and defining
their relationships.

This does add a bit of boilerplate but it adds better structure to the
models.

This also forces us to make all containers accessible from the outside
of the models, which is probably a good thing to make things easier
going forward.