Move Provider Training website into its own Software System
As of 1st November 2020, the CCMS Training website has been
replaced with a more generic learning platform.

Announcement:
```
A new website is being launched on 1 November 2020 to replace the existing Client and Cost Management System (CCMS) training website. Resources will be gradually transferred to the new website. User improvements include:

...

Long-term, the aim will be for the new training and support website to host crime as well as civil training information. A redirect will automatically take you to the new home page from 1 November if you click on the old CCMS training URL.
```

Closes #43